pyproject.toml
[project]
name = "curso-dunossauro"
version = "0.1.0"
description = ""
authors = [
    {name = "lbauch",email = "bauchspiess.lucas@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.129.0,<0.130.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "pydantic-settings (>=2.13.0,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "tzdata (>=2025.3,<2026.0)",
    "psycopg[binary] (>=3.3.3,<4.0.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.15.1,<0.16.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "freezegun (>=1.5.5,<2.0.0)",
    "testcontainers (>=4.14.1,<5.0.0)"
]

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']

[tool.ruff.format]
quote-style = "single"

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
async_default_fixture_loop_scope = 'function'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
pre_run_dev = 'task format'
run_dev = 'fastapi dev curso_dunossauro/app.py'
run_prod = 'fastapi curso_dunossauro/app.py --host 0.0.0.0'
pre_test = 'task format'
test = 'pytest -s --maxfail=5 --cov=curso_dunossauro -vv'
post_test = 'coverage html'
upgrade_db = 'alembic revision --autogenerate -m "Atualizando db"'
post_upgrade_db = 'alembic upgrade head'

[tool.coverage.run]
concurrency = ["thread", "greenlet"]